requirements.txt
tls_client

# --- HTML/XML Parsing ---
beautifulsoup4==4.12.3
lxml==5.3.0

# --- Telegram Bot API ---
pytelegrambotapi==4.15.2

# --- System & Utilities ---
psutil==6.0.0
colorama==0.4.6

# --- Security & Hashing ---
passlib==1.7.4
bcrypt==4.1.3

# --- Image Processing ---
Pillow==10.4.0

# --- AI & Data Models ---
# Внимание: g4f часто ломается из-за изменений на сайтах нейросетей. 
# Если AI не отвечает, попробуйте обновить эту библиотеку до последней (убрав ==...).
g4f==0.3.3.1
pydantic==2.8.2

# --- Database, Network & Utils ---
httpx==0.24.1
tonutils==0.1.5
pymysql==1.1.1
redis==5.0.8
requests-toolbelt==1.0.0

# --- Data Visualization ---
matplotlib==3.7.1

# --- Compatibility ---
# Удерживаем версию ниже 2.0 для совместимости с некоторыми старыми либами
urllib3==1.26.19
pycryptodome